requirements.txt
beautifulsoup4
crewai
fastapi
google-generativeai
huggingface-hub
matplotlib
onnxruntime
pandas
pdfplumber
pinecone
pillow
playwright
pydantic
python-dotenv
python-multipart
requests
scipy
seaborn
uvicorn
uvloop
pdfplumber 
mammoth 
python-docx
langchain
langchain-core
langchain-experimental
langchain-google-genai